requirements.txt
# Django==4.0.4; python_version >= '3.8'
# Django==3.2.14; python_version < '3.8'
# gunicorn==20.1.0
# psycopg2-binary==2.9.3
# django-environ==0.9.0
# google-cloud-secret-manager==2.11.1
# django-storages[google]==1.12.3
Django==3.2.14 
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-environ==0.9.0
google-cloud-secret-manager==2.11.1
django-storages[google]==1.12.3
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
django-ckeditor==6.0.0
